Use new enum class (strong enum) support. Not 100% - but most places in headers.
Next step will be to change naming convention for names (with strong named enums, no point
in the 'e' prefix anymore).

And had to deal with explict casts to int as needed.

And fixed BUG (untested yet) in ThreadPool::AbortTask () - due to LAX old enum stuff (not a!=b).

Still not tested on GCC/Unix yet.