Major change to socket code - mimicing the refactoring just done to Streams (Ptr).
To make these classes more accurately reflect the fact that they operate as SmartPtrs -
Socket, ConnectionOrientedSocket, ConnectionlessSocket and ConnectionOrientedMasterSocket
have ALL been updated to no longer be default constructible, and no longer be copyable/movable.

***NOT BACKWARD COMPATIBLE***

But - this is easy to adopt to. There are nested 'Ptr' classes inside each class, which are just like
the owning class (e.g. Socket) except for being copyable and default constructible.

So if you have code like:
	XXX (where XXX is Socket, or ConnectionOrientedSocket for example), just replace it with
	XXX::Ptr everywhere you have a copyable object (only exception being where you construct
	a specific kind of socket).

So
	ConnectionOrientedSocket       s  = localSocketToAcceptOn.Accept ();
becomes:
      ConnectionOrientedSocket::Ptr s  = localSocketToAcceptOn.Accept ();